Split review extension into runtime modules
