+++ .gitattributes
@@ -65,6 +65,9 @@ extensions/the-last-harness/primary-agent-runtime.js linguist-generated=true
 extensions/the-last-harness/profile-state.js linguist-generated=true
 extensions/the-last-harness/prompts.js linguist-generated=true
 extensions/the-last-harness/resources.js linguist-generated=true
+extensions/the-last-harness/review-args.js linguist-generated=true
+extensions/the-last-harness/review-envelope.js linguist-generated=true
+extensions/the-last-harness/review-github.js linguist-generated=true
 extensions/the-last-harness/review.js linguist-generated=true
 extensions/the-last-harness/shell-parser.js linguist-generated=true
 extensions/the-last-harness/startup-tip.js linguist-generated=true
