+++ extensions/the-last-harness/review.js
@@ -1,12 +1,16 @@
-import { lstat, open, readdir, readFile } from "node:fs/promises";
-import { join, relative, resolve } from "node:path";
+import { lstat, readdir } from "node:fs/promises";
+import { join, resolve } from "node:path";
+import { REVIEW_MODES, REVIEW_MODE_DESCRIPTIONS, decideBranchAction, tokenizeArgs, parseReviewArgs } from "./review-args.js";
+export { REVIEW_MODES, decideBranchAction, parseReviewArgs } from "./review-args.js";
+import { buildReviewEnvelope, parseNullDelimitedGitPaths, buildSnapshotParts, appendUntrackedSnapshot, } from "./review-envelope.js";
+export { buildReviewEnvelope } from "./review-envelope.js";
+import { isGhGraphqlQuotaFailure, resolveGitHubPrRef, fetchPrMetadataViaRest, fetchPrDiffViaRest, } from "./review-github.js";
 import { DynamicBorder, getAgentDir, getSelectListTheme, SettingsManager } from "@earendil-works/pi-coding-agent";
 import { Container, matchesKey, SelectList, Text } from "@earendil-works/pi-tui";
 import { primaryAgentSelectionFromBranch, resolvePrimaryAgentConfig } from "../the-last-harness-primary-agent.mjs";
 const REVIEW_TITLE = "Choose a review mode";
 const REVIEW_PICKER_HINT = "↑/↓ to move  Enter to confirm  Esc to cancel";
 const REVIEW_DEFAULT_BRANCH_BASE = "main";
-const REVIEW_PICKER_ONLY_GUIDANCE = "/review is picker-only. Run /review with no arguments, then choose a mode in the picker. Typed shortcuts like `/review pr 123` and `--extra` are no longer supported.";
 const REVIEW_TUI_REQUIRED_MESSAGE = "/review requires the interactive TUI picker. Re-run /review in the TLH UI.";
 const REVIEW_REQUIRED_PRIMARY = "architect";
 function isRecord(value) {
@@ -29,101 +33,6 @@ function currentReviewPrimaryAgentSelection(ctx) {
 }
 function reviewPrimaryBlockedMessage(activePrimary) {
     return `/review only works while the architect primary agent is active. Current primary agent: ${activePrimary}. Switch to architect with /switch-primary-agent architect (or Shift+Tab), then rerun /review.`;
-}
-export const REVIEW_MODES = ["uncommitted", "branch", "commit", "pr", "folder"];
-const REVIEW_MODE_DESCRIPTIONS = {
-    uncommitted: "Review staged/unstaged changes plus untracked non-gitignored files",
-    branch: "Review commits on the current branch vs a chosen base (prompted; blank defaults to main)",
-    commit: "Review a single commit by SHA",
-    pr: "Review a pull request by number or URL",
-    folder: "Review files in one or more folders",
-};
-const REVIEW_UNTRACKED_BEGIN_DELIMITER = "--- begin untracked files ---";
-const REVIEW_UNTRACKED_END_DELIMITER = "--- end untracked files ---";
-export function decideBranchAction(params) {
-    const { currentBranch, prHead, isDirty, userConfirm } = params;
-    if (currentBranch === prHead)
-        return "proceed";
-    if (isDirty)
-        return "abort-dirty";
-    return userConfirm ? "switch" : "abort-cancelled";
-}
-function tokenizeArgs(raw) {
-    if (!raw.trim())
-        return [];
-    const tokens = [];
-    let current = "";
-    let inSingleQuote = false;
-    let inDoubleQuote = false;
-    for (const ch of raw) {
-        if (ch === "'" && !inDoubleQuote) {
-            inSingleQuote = !inSingleQuote;
-        }
-        else if (ch === '"' && !inSingleQuote) {
-            inDoubleQuote = !inDoubleQuote;
-        }
-        else if (/\s/.test(ch) && !inSingleQuote && !inDoubleQuote) {
-            if (current) {
-                tokens.push(current);
-                current = "";
-            }
-        }
-        else {
-            current += ch;
-        }
-    }
-    if (current) {
-        tokens.push(current);
-    }
-    return tokens;
-}
-export function parseReviewArgs(argv) {
-    if (argv.length === 0) {
-        return { pickerRequested: true };
-    }
-    return {
-        pickerRequested: false,
-        message: REVIEW_PICKER_ONLY_GUIDANCE,
-    };
-}
-export function buildReviewEnvelope(parsed, ctx) {
-    const { mode, extra } = parsed;
-    const lines = [];
-    lines.push("[/review]");
-    lines.push(`mode: ${mode}`);
-    if (parsed.mode === "branch" && parsed.base) {
-        lines.push(`base: ${parsed.base}`);
-    }
-    else if (parsed.mode === "commit" && parsed.sha) {
-        lines.push(`sha: ${parsed.sha}`);
-    }
-    else if (parsed.mode === "pr" && parsed.nOrUrl) {
-        lines.push(`pr: ${parsed.nOrUrl}`);
-    }
-    else if (parsed.mode === "folder" && parsed.paths.length > 0) {
-        lines.push(`paths: ${parsed.paths.join(" ")}`);
-    }
-    if (ctx?.currentBranch !== undefined) {
-        lines.push(`current-branch: ${ctx.currentBranch}`);
-    }
-    if (ctx?.checkout?.performed) {
-        lines.push(`checkout: switched-from ${ctx.checkout.priorBranch}`);
-        lines.push(`note: previously on ${ctx.checkout.priorBranch}; run \`git checkout -\` to return.`);
-    }
-    if (extra === undefined) {
-        lines.push("extra: (none)");
-    }
-    else {
-        lines.push("extra:");
-        lines.push(extra);
-    }
-    const hasBody = ctx?.body !== undefined;
-    const fenceKind = hasBody ? (ctx?.bodyKind ?? "diff") : "(pending)";
-    const bodyText = hasBody ? escapeEnvelopeFenceLines(ctx?.body, fenceKind) : "(no body gathered)";
-    lines.push(`--- begin ${fenceKind} ---`);
-    lines.push(bodyText);
-    lines.push(`--- end ${fenceKind} ---`);
-    return lines.join("\n");
 }
 function makePickedArgs(mode) {
     return { mode, extra: undefined };
@@ -219,155 +128,6 @@ function rejectFlagLike(value, fieldName) {
         return { ok: false, message: `${fieldName} cannot start with '-' (got '${value}'). If this is intentional, run the underlying command manually.` };
     }
     return { ok: true };
-}
-function isGhGraphqlQuotaFailure(stderr) {
-    return /graphql/i.test(stderr) && /(rate limit|quota|submitted too quickly)/i.test(stderr);
-}
-function parseGitHubPrUrl(value) {
-    try {
-        const url = new URL(value);
-        if (url.hostname !== "github.com") {
-            return undefined;
-        }
-        const match = url.pathname.match(/^\/([^/]+)\/([^/]+)\/pull\/(\d+)(?:\/.*)?$/);
-        if (!match) {
-            return undefined;
-        }
-        return { owner: match[1], repo: match[2], number: Number.parseInt(match[3], 10) };
-    }
-    catch {
-        return undefined;
-    }
-}
-function parseGitHubRepoSlug(value) {
-    const match = value.trim().match(/^([^/\s]+)\/([^/\s]+)$/u);
-    if (!match) {
-        return undefined;
-    }
-    return { owner: match[1], repo: match[2] };
-}
-function parseGitHubRemoteUrl(value) {
-    const trimmed = value.trim();
-    const sshMatch = trimmed.match(/^git@github\.com:([^/]+)\/([^/]+?)(?:\.git)?$/i);
-    if (sshMatch) {
-        return { owner: sshMatch[1], repo: sshMatch[2] };
-    }
-    try {
-        const url = new URL(trimmed);
-        if (url.hostname !== "github.com") {
-            return undefined;
-        }
-        const match = url.pathname.match(/^\/([^/]+)\/([^/]+?)(?:\.git)?\/?$/);
-        if (!match) {
-            return undefined;
-        }
-        return { owner: match[1], repo: match[2] };
-    }
-    catch {
-        return undefined;
-    }
-}
-async function resolveGitHubRepoRefFromGhDefault(pi, cwd) {
-    const defaultRepoResult = await pi.exec("gh", ["repo", "set-default", "--view"], { cwd });
-    if (defaultRepoResult.code !== 0) {
-        return undefined;
-    }
-    for (const line of defaultRepoResult.stdout.split(/\r?\n/u)) {
-        const repoRef = parseGitHubRepoSlug(line);
-        if (repoRef) {
-            return repoRef;
-        }
-    }
-    return undefined;
-}
-async function resolveGitHubRepoRefFromLocalRemotes(pi, cwd) {
-    const remoteListResult = await pi.exec("git", ["remote"], { cwd });
-    if (remoteListResult.code !== 0) {
-        const firstLine = remoteListResult.stderr.split("\n")[0]?.trim() || "git remote failed";
-        return { ok: false, message: `could not list git remotes: ${firstLine}` };
-    }
-    const remoteNames = remoteListResult.stdout
-        .split(/\r?\n/u)
-        .map((name) => name.trim())
-        .filter(Boolean);
-    if (remoteNames.length === 0) {
-        return { ok: false, message: "could not resolve GitHub repository because this repo has no git remotes" };
-    }
-    const orderedRemoteNames = Array.from(new Set(["origin", ...remoteNames]));
-    const remoteFailures = [];
-    for (const remoteName of orderedRemoteNames) {
-        const remoteUrlResult = await pi.exec("git", ["remote", "get-url", remoteName], { cwd });
-        if (remoteUrlResult.code !== 0) {
-            const firstLine = remoteUrlResult.stderr.split("\n")[0]?.trim() || `git remote get-url ${remoteName} failed`;
-            remoteFailures.push(`${remoteName}: ${firstLine}`);
-            continue;
-        }
-        const repoRef = parseGitHubRemoteUrl(remoteUrlResult.stdout);
-        if (repoRef) {
-            return { ok: true, repoRef };
-        }
-        remoteFailures.push(`${remoteName}: unsupported remote URL '${remoteUrlResult.stdout.trim()}'`);
-    }
-    return {
-        ok: false,
-        message: `could not parse a GitHub owner/repo from local git remotes (${remoteFailures.join("; ")})`,
-    };
-}
-async function resolveGitHubPrRef(pi, cwd, nOrUrl, prNumberHint) {
-    const urlRef = parseGitHubPrUrl(nOrUrl);
-    if (urlRef) {
-        return { ok: true, prRef: urlRef };
-    }
-    if (prNumberHint === undefined) {
-        return { ok: false, message: `could not resolve a PR number from '${nOrUrl}'` };
-    }
-    const ghDefaultRepoRef = await resolveGitHubRepoRefFromGhDefault(pi, cwd);
-    if (ghDefaultRepoRef) {
-        return { ok: true, prRef: { ...ghDefaultRepoRef, number: prNumberHint } };
-    }
-    const repoRefResult = await resolveGitHubRepoRefFromLocalRemotes(pi, cwd);
-    if (repoRefResult.ok === false) {
-        return { ok: false, message: repoRefResult.message };
-    }
-    return { ok: true, prRef: { ...repoRefResult.repoRef, number: prNumberHint } };
-}
-async function fetchPrMetadataViaRest(pi, cwd, prRef) {
-    const result = await pi.exec("gh", ["api", `repos/${prRef.owner}/${prRef.repo}/pulls/${prRef.number}`], { cwd });
-    if (result.code !== 0) {
-        const firstLine = result.stderr.split("\n")[0]?.trim() || "gh api failed";
-        return { ok: false, message: firstLine };
-    }
-    try {
-        const payload = JSON.parse(result.stdout);
-        const number = typeof payload.number === "number" ? payload.number : prRef.number;
-        const headRefName = payload.head?.ref;
-        const baseRefName = payload.base?.ref;
-        if (!headRefName || !baseRefName) {
-            return { ok: false, message: "REST PR metadata response was missing head/base refs" };
-        }
-        return {
-            ok: true,
-            prData: {
-                number,
-                headRefName,
-                baseRefName,
-                isCrossRepository: typeof payload.head?.repo?.full_name === "string" && typeof payload.base?.repo?.full_name === "string"
-                    ? payload.head.repo.full_name !== payload.base.repo.full_name
-                    : false,
-            },
-        };
-    }
-    catch {
-        return { ok: false, message: "Could not parse REST PR metadata response" };
-    }
-}
-async function fetchPrDiffViaRest(pi, cwd, prRef) {
-    const result = await pi.exec("gh", ["api", "-H", "Accept: application/vnd.github.v3.diff", `repos/${prRef.owner}/${prRef.repo}/pulls/${prRef.number}`], { cwd });
-    if (result.code !== 0) {
-        const firstLine = result.stderr.split("\n")[0]?.trim() || "gh api failed";
-        return { ok: false, message: firstLine };
-    }
-    return { ok: true, diff: result.stdout };
 }
 async function gatherUncommitted(pi, cmdCtx) {
     const cwd = cmdCtx.cwd;
@@ -486,18 +246,6 @@ async function gatherCommit(pi, cmdCtx, sha) {
         ctx: { currentBranch, body: showResult.stdout, bodyKind: "diff" },
     };
 }
-async function isBinaryFile(filePath) {
-    let handle;
-    try {
-        handle = await open(filePath, "r");
-        const buf = Buffer.alloc(8192);
-        const { bytesRead } = await handle.read(buf, 0, 8192, 0);
-        return buf.subarray(0, bytesRead).includes(0);
-    }
-    finally {
-        await handle?.close();
-    }
-}
 async function walkDir(dir) {
     const entries = await readdir(dir, { withFileTypes: true });
     const files = [];
@@ -514,101 +262,6 @@ async function walkDir(dir) {
         }
     }
     return files;
-}
-function parseNullDelimitedGitPaths(stdout) {
-    return stdout.split("\0").filter((filePath) => filePath.length > 0);
-}
-function escapeDelimitedContentLine(line) {
-    return `\\${line}`;
-}
-function escapeContentDelimiters(content) {
-    return content
-        .split("\n")
-        .map((line) => {
-        if (line === "--- begin snapshot ---"
-            || line === "--- end snapshot ---"
-            || line === REVIEW_UNTRACKED_BEGIN_DELIMITER
-            || line === REVIEW_UNTRACKED_END_DELIMITER
-            || /^--- (?:file|untracked file): .* ---$/.test(line)) {
-            return escapeDelimitedContentLine(line);
-        }
-        return line;
-    })
-        .join("\n");
-}
-function escapeEnvelopeFenceLines(body, fenceKind) {
-    const beginFence = `--- begin ${fenceKind} ---`;
-    const endFence = `--- end ${fenceKind} ---`;
-    return body
-        .split("\n")
-        .map((line) => (line === beginFence || line === endFence ? escapeDelimitedContentLine(line) : line))
-        .join("\n");
-}
-function renderDelimitedPath(relPath) {
-    return JSON.stringify(relPath)
-        .replace(/[\u007f-\u009f\u2028\u2029]/g, (ch) => `\\u${ch.charCodeAt(0).toString(16).padStart(4, "0")}`)
-        .replace(/\[/g, "\\u005b")
-        .replace(/\]/g, "\\u005d");
-}
-function getNonRegularSnapshotMarker(relPath, pathStat) {
-    const renderedPath = renderDelimitedPath(relPath);
-    if (pathStat.isSymbolicLink()) {
-        return `[skipped symlink: ${renderedPath}]`;
-    }
-    if (pathStat.isDirectory()) {
-        return `[skipped directory: ${renderedPath}]`;
-    }
-    if (!pathStat.isFile()) {
-        return `[skipped non-regular entry: ${renderedPath}]`;
-    }
-    return undefined;
-}
-async function buildSnapshotParts(cwd, filePaths, label) {
-    const parts = [];
-    for (const filePath of filePaths) {
-        const relPath = relative(cwd, filePath);
-        const renderedPath = renderDelimitedPath(relPath);
-        let pathStat;
-        try {
-            pathStat = await lstat(filePath);
-        }
-        catch {
-            parts.push(`[skipped lstat failure: ${renderedPath}]`);
-            continue;
-        }
-        const nonRegularMarker = getNonRegularSnapshotMarker(relPath, pathStat);
-        if (nonRegularMarker) {
-            parts.push(nonRegularMarker);
-            continue;
-        }
-        let bin;
-        try {
-            bin = await isBinaryFile(filePath);
-        }
-        catch {
-            parts.push(`[skipped binary detection failure: ${renderedPath}]`);
-            continue;
-        }
-        if (bin) {
-            parts.push(`[skipped binary: ${renderedPath}]`);
-            continue;
-        }
-        try {
-            const content = escapeContentDelimiters(await readFile(filePath, "utf8"));
-            parts.push(`--- ${label}: ${renderedPath} ---\n${content}`);
-        }
-        catch {
-            parts.push(`[skipped read failure: ${renderedPath}]`);
-        }
-    }
-    return parts;
-}
-function appendUntrackedSnapshot(diffBody, untrackedParts) {
-    if (untrackedParts.length === 0) {
-        return diffBody;
-    }
-    const untrackedBody = [REVIEW_UNTRACKED_BEGIN_DELIMITER, ...untrackedParts, REVIEW_UNTRACKED_END_DELIMITER].join("\n");
-    return diffBody.trim().length > 0 ? `${diffBody}\n\n${untrackedBody}` : untrackedBody;
 }
 async function gatherFolder(pi, cmdCtx, paths) {
     if (paths.length === 0) {
