+++ tests/package-runtime-smoke.test.mjs
@@ -93,7 +93,7 @@ test("packed TLH generated JavaScript loads and reloads through pinned Pi 0.80.6
 		.filter((path) => !path.endsWith(".d.ts"))
 		.map((path) => path.replace(/\.ts$/, ".js"))
 		.sort();
-	assert.equal(generatedExtensionPaths.length, 63);
+	assert.equal(generatedExtensionPaths.length, 66);
 	for (const generatedPath of generatedExtensionPaths) {
 		assert.ok(packedPaths.has(generatedPath), `npm pack omitted generated extension module ${generatedPath}`);
 	}
